--- infa3/infa3/__init__.py
@@ -1,7 +1,0 @@
-# pmrep interface
-from infa3.pmrep import Pmrep
-
-# exceptions
-from infa3.exceptions import InfaError, InfaPmrepError
-
-__all__ = ['Pmrep', 'InfaError', 'InfaPmrepError']
