--- infa3/infa3/exceptions.py
@@ -1,17 +1,0 @@
-"""
-    infa.exceptions
-    ~~~~~~~~~~~~~~~
-
-    Infa exceptions.
-
-"""
-
-class InfaError(Exception):
-    """Baseclass for all Infa errors."""
-
-    pass
-
-class InfaPmrepError(InfaError):
-    """Raised if a pmrep error occurs."""
-
-    pass
